feat: ab managed portal
