+++ docs/src/reference/portal/config.md
@@ -5,7 +5,12 @@ Portal config lives under `[portal]` in `~/.agent-box.toml`.
 ## Top-level keys
 
 - `enabled` (bool, default: `true`)
+- `global` (bool, default: `false`)
+  - `true`: use a long-lived user-managed `agent-portal-host`
+  - `false`: `ab spawn` starts a per-container `agent-portal-host` and tears it down when the container exits
 - `socket_path` (string, default: `/run/user/<uid>/agent-portal/portal.sock`)
+  - used directly when `global = true`
+  - ignored by `ab spawn` when `global = false`, because `ab` allocates a unique per-container socket path
 - `prompt_command` (string|null, default: unset)
 - `timeouts.request_ms` (u64, default: `0` = no timeout)
 - `timeouts.prompt_ms` (u64, default: `0` = no timeout)
@@ -15,6 +20,41 @@ Portal config lives under `[portal]` in `~/.agent-box.toml`.
 - `limits.rate_burst` (u32, default: `10`)
 - `limits.max_clipboard_bytes` (usize, default: `20971520`)
 - `clipboard.allowed_mime` (array of strings, default: `image/png`, `image/jpeg`, `image/webp`)
+
+## Operation modes
+
+Managed per-container mode is the default.
+
+### Global mode
+
+Use this when you want to run Portal yourself, for example via a user service.
+
+```toml
+[portal]
+enabled = true
+global = true
+socket_path = "/run/user/1000/agent-portal/portal.sock"
+```
+
+You only need to set `global = true` when you want to override the default managed mode.
+
+### Managed per-container mode
+
+Use this when you want `ab` to start a dedicated Portal instance for each `ab spawn`.
+
+```toml
+[portal]
+enabled = true
+global = false
+```
+
+In this mode, `ab`:
+
+1. starts an in-process Portal host before launching the container
+2. chooses a unique socket path for that spawn
+3. mounts that socket into the container
+4. sets `AGENT_PORTAL_SOCKET` for the container
+5. shuts the Portal host down after the container exits
 
 ## Policy defaults
 
@@ -46,6 +86,7 @@ gh_exec = "ask_for_all"
 ```toml
 [portal]
 enabled = true
+global = true
 socket_path = "/run/user/1000/agent-portal/portal.sock"
 prompt_command = "rofi -dmenu -p 'agent-portal'"
